Fix Win64 compile error caused by r12272.
R=mstarzinger@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10830213

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@12276 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/extensions/statistics-extension.cc
@@ -51,9 +51,10 @@ static void AddCounter(v8::Local<v8::Object> object,
 }
 
 static void AddNumber(v8::Local<v8::Object> object,
-                      double value,
+                      intptr_t value,
                       const char* name) {
-  object->Set(v8::String::New(name), v8::Number::New(value));
+  object->Set(v8::String::New(name),
+              v8::Number::New(static_cast<double>(value)));
 }
